backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlmodel==0.0.14
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
email-validator==2.1.0
python-dotenv==1.0.0
pydantic-settings==2.1.0
pytest==7.4.3
httpx==0.25.2
